+++ components/CardPostRecommend.vue
@@ -1,11 +1,11 @@
 <template>
-  <div class="card _card">
-    <div class="card-image _card-image">
+  <div class="media _media-recommend">
+    <div class="media-left">
       <nuxt-link :to="localePath(`/posts/${postUrl}`)">
-        <figure class="image is-4by3">
+        <figure class="image is-96x96">
           <client-only>
             <img
-              class="_thumbnail-image"
+              class="_thumbnail-image-recommend"
               v-lazy="postThumbnail"
               style="display: none"
               onload="this.style.display = 'block'"
@@ -14,37 +14,33 @@
           </client-only>
         </figure>
       </nuxt-link>
-      <figure class="_card-avatar">
-        <client-only>
-          <img
-            class="_creator-avatar"
-            v-lazy="creatorAvatar"
-            style="display: none"
-            onload="this.style.display = 'block'"
-            alt="creator_avatar"
-          />
-        </client-only>
-      </figure>
     </div>
-    <div class="card-content _card-content">
+    <div class="media-content" style="overflow: hidden;">
       <div class="content">
-        <div class="media">
-          <div class="media-content">
-            <p class="title is-4">{{ postTitle }}</p>
-          </div>
-        </div>
+        <strong>
+          <nuxt-link class="is-size-6" :to="localePath(`/posts/${postUrl}`)">
+            {{ postTitle }}
+          </nuxt-link>
+        </strong>
+      </div>
+      <div class="tags">
+        <span class="tag is-info is-small">{{ postCategory }}</span>
+        <span
+          v-for="(tag, index) in postTags"
+          :key="index"
+          class="tag is-warning is-small"
+          >{{ tag }}</span
+        >
       </div>
     </div>
   </div>
 </template>
 
 <script>
-import { mapGetters } from "vuex";
-import { fetchDesc } from "~/libs/helpers";
 import { categories } from "~/libs/lists";
 
 export default {
-  name: "v-card-post-recommend",
+  name: "v-card-post-horizontal",
   props: {
     value: {
       type: Object,
@@ -58,9 +54,6 @@ export default {
     postTitle() {
       return this.value.title;
     },
-    postMode() {
-      return this.value.mode
-    },
     postCategory() {
       const category = categories.find(item => item.id === this.value.category);
       return category.name;
@@ -68,28 +61,12 @@ export default {
     postTags() {
       return this.value.tags;
     },
-    postDescription() {
-      return fetchDesc(this.value.html);
-    },
-    postUpdatedDate() {
-      return this.value.updatedDate;
-    },
     postThumbnail() {
       if (this.value.images) {
         return this.value.images[0].url;
       } else {
         return "/icon-photo.png";
       }
-    },
-    creatorAvatar() {
-      if (this.value.creator.avatar) {
-        return this.value.creator.avatar.url;
-      } else {
-        return "/icon-user.png";
-      }
-    },
-    creatorUsername() {
-      return this.value.creator.username;
     }
   }
 };